feat: Adding Home component
Home component will be the entry view for the application it will recieve data from the main component
and render nested elements based on that.

Home component will list 3 shelf areas (Want To Read, Currently Reading, Read) in which book components will be categorized.
It will also contain a Link to navigate to the route /search were the Search component will be rendered